Reestructuración de carpetas de todo el proyecto
