+++ src/assets/styles/calendar/Calendar.scss
@@ -69,8 +69,8 @@
     overflow: hidden;
 
     & .imagen-home-calendario{
-        width: 370px;
-        height: 370px;
+        width: 340px;
+        height: 340px;
         position: absolute;
         right: 0;
         bottom: 0;
