+++ src/components/notifications/notificationsAlerts/DeleteKr.js
