+++ src/components/structure/Navbar.js
@@ -1,16 +1,63 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { MdNotifications, MdInput } from "react-icons/md";
+import { MdInput } from "react-icons/md";
 import "../../assets/styles/structure/Navbar.css";
 import { IconContext } from "react-icons";
 import Logo from "./Logo";
-import foto from "../../assets/fotoTemporal.jpg";
+import foto from "../../assets/img/fotoTemporal.jpg";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { auth } from '../../functions/firebaseAuth';
-import { SingOut } from './SignOut'
-
+import { SingOut } from './SignOut';
+import Bell from "../notifications/modalbell/ModalBell";
 
 function Navbar() {
+  const state = {
+    data: [
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3a',
+            name: 'Notification Proyect',
+            typeN: 'OKR',
+            detail: '30',
+            viewed: 'false'
+        },
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3b',
+            name: 'Notification Proyect',
+            typeN: 'OKR',
+            detail: '70',
+            viewed: 'true'
+        },
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3b',
+            name: 'Notification Proyect',
+            typeN: 'OKR',
+            detail: '100 Or Finished',
+            viewed: 'true'
+        },
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3a',
+            name: 'Notification Proyect',
+            typeN: 'KR',
+            detail: '30',
+            viewed: 'true'
+        },
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3b',
+            name: 'Notification Proyect',
+            typeN: 'KR',
+            detail: '70',
+            viewed: 'true'
+        },
+        {
+            id: 'd00d3614-101a-44ca-b6c2-0be075aeed3b',
+            name: 'Notification Proyect ',
+            typeN: 'KR',
+            detail: '100',
+            viewed: 'true'
+        },
+    ],
+};
+
   const [user] = useAuthState(auth);
   return (
     <>
@@ -25,7 +72,7 @@ function Navbar() {
             <ul id = "list-notifications">
               <li id = "notifications-icon" className="notifications-icon">
                 <span id = "span-notificaciones-navbar">
-                  <MdNotifications />
+                <li><Bell state={state.data} /></li>
                 </span>
               </li>
               <li>
