--- src/pages/Home.js
@@ -1,19 +1,0 @@
-import React from 'react'
-import EditKr from '../notifications/EditKr'
-import EditOkr from '../notifications/EditOkr'
-import DeleteOkr from '../notifications/DeleteOkr'
-import DeleteKr from '../notifications/DeleteKr'
-
-export default function Home() {
-    return (
-        <div className='home'>
-            <h1>Home</h1>
-            {/* <button onClick={EditOkr} data-testid="edit-okr-notification"> Editar OKR </button>
-            <button onClick={EditKr} data-testid="edit-kr-notification"> Editar KR </button>
-            <button onClick={DeleteOkr} data-testid="delete-okr-notification"> Eliminar OKR </button>
-            <button onClick={DeleteKr} data-testid="delete-kr-notification"> Eliminar KR </button> */}
-        </div>
-
-
-    )
-}
