+++ src/pages/administration/Support.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Navbar from '../../components/structure/Navbar';
+import Sidebar from '../../components/structure/Sidebar';
+
+export default function Support() {
+    return (
+        <div>
+            <Navbar/>
+            <Sidebar/>
+            <h1>Support</h1>
+        </div>
+    )
+}
