+++ src/pages/confignotification/ConfigNotification.js
@@ -1,5 +1,4 @@
 import React from 'react'
-
 import '../../assets/styles/notifications/Notification.scss'
 import Button from '@material-ui/core/Button';
 import swal from 'sweetalert';
